chore : 🚚 의도가 명확하게 들어나도록 이름 변경 함수의 이름이 함수가 가진 역할을 명확하게 나타낼 수 있도록 rename 작업 가지고 console.log의 경우 삭제해 주었다. index.js / createCarInstance.js / computeGameResult.js / drawCarRacing.js / drawRacingCarGameWinner.js / hideRacingCountAndResultElements.js / playRacingCarGame.js / showRacingResultTitleElement.js
