+++ src/index.js
@@ -1,9 +1,9 @@
 import { $ } from './dom/dom.js';
+import hideRacingCountAndResultElements from './modules/hideRacingCountAndResultElements.js';
 import getUserCarNamesInput from './modules/getUserCarNamesInput.js';
-import hideRacingCountShow from './modules/hideRacingCountShow.js';
 import showRacingCountInput from './modules/showRacingCountInput.js';
 import getUserRacingCountInput from './modules/getUserRacingCountInput.js';
-import showRacingResultTitle from './modules/showRacingResultTitle.js';
+import showRacingResultTitleElement from './modules/showRacingResultTitleElement.js';
 import createCarInstance from './car/createCarInstance.js';
 import playRacingCarGame from './modules/playRacingCarGame.js';
 
@@ -16,11 +16,12 @@ export default function RacingcarGame() {
 
   this.init = () => {
     initEventListener();
-    hideRacingCountShow();
+    hideRacingCountAndResultElements();
   };
 
-  const gameStart = () => {
+  const handleCarNamesEvent = () => {
     this.racingInfoObject.carNames = getUserCarNamesInput();
+    //유저가 입력한 자동차이름이 유효한 경우에만 다음 단계의 이벤트가 실행된다
     if (this.racingInfoObject.carNames !== false) {
       showRacingCountInput();
       addRacingCountEventListener();
@@ -28,13 +29,14 @@ export default function RacingcarGame() {
     return;
   };
 
-  const gameContinue = () => {
+  const handleRacingCountEvent = () => {
     this.racingInfoObject.gameCount = getUserRacingCountInput();
+    //유저가 입력한 racing 횟수가 유효한 경우에만 다음 단계의 이벤트가 실행된다
     if (this.racingInfoObject.gameCount !== false) {
       this.racingInfoObject.carInstanceList = createCarInstance(
         this.racingInfoObject.carNames
       );
-      showRacingResultTitle();
+      showRacingResultTitleElement();
       playRacingCarGame(this.racingInfoObject);
     }
   };
@@ -43,14 +45,14 @@ export default function RacingcarGame() {
     $('#racing-count-form').addEventListener('submit', (e) => {
       e.preventDefalut();
     });
-    $('#racing-count-submit').addEventListener('click', gameContinue);
+    $('#racing-count-submit').addEventListener('click', handleRacingCountEvent);
   };
 
   const initEventListener = () => {
     $('#car-names-form').addEventListener('submit', (e) => {
       e.preventDefalut();
     });
-    $('#car-names-submit').addEventListener('click', gameStart);
+    $('#car-names-submit').addEventListener('click', handleCarNamesEvent);
   };
 }
