+++ src/modules/computeGameResult.js
@@ -2,7 +2,6 @@ import createRandomNumber from './createRandomNumber.js';
 
 function decideStoporFoward() {
   const num = createRandomNumber();
-  console.log(num);
   if (num >= 4) {
     return true;
   } else if (num < 4) {
@@ -10,7 +9,7 @@ function decideStoporFoward() {
   }
 }
 
-export default function playRacingCarGame(carInstanceList) {
+export default function computeGameResult(carInstanceList) {
   carInstanceList.forEach((item) => {
     if (decideStoporFoward()) {
       item.plusCount();
