+++ src/modules/hideRacingCountAndResultElements.js
@@ -1,6 +1,6 @@
 import { $ } from '../dom/dom.js';
 
-export default function hideRacingCountShow() {
+export default function hideRacingCountAndResultElements() {
   $('#racing-count-title').style.display = 'none';
   $('#racing-count-form').style.display = 'none';
   $('#racing-result-title').style.display = 'none';
