+++ src/modules/playRacingCarGame.js
@@ -1,11 +1,11 @@
-import drawCarRacing from '../car/drawCarRacing.js';
+import drawCarRacing from './drawCarRacing.js';
 import computeGameResult from './computeGameResult.js';
-import drawRacingCarGameResult from './drawRacingCarGameResult.js';
+import drawRacingCarGameWinner from './drawRacingCarGameWinner.js';
 
 export default function playRacingCarGame({ gameCount, carInstanceList }) {
   for (let i = 0; i < gameCount; i++) {
     computeGameResult(carInstanceList);
     drawCarRacing(carInstanceList);
   }
-  drawRacingCarGameResult(carInstanceList);
+  drawRacingCarGameWinner(carInstanceList);
 }
