+++ src/modules/showRacingResultTitleElement.js
@@ -1,4 +1,4 @@
 import { $ } from '../dom/dom.js';
-export default function showRacingResultTitle() {
+export default function showRacingResultTitleElement() {
   $('#racing-result-title').style.display = 'block';
 }
